refactor: - Modificando a forma como o BiSmart Cliente funciona. - Apenas configurador agora - run.bat para rodar quando o computador liga
