+++ auto_run/run.bat
@@ -1,0 +1,1 @@
+node script.js
