--- config.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Get data</title>
-  </head>
-  <style>
-    body {
-      margin-left: 30px;
-    }
-    p {
-      margin: 0px;
-    }
-  </style>
-  <body>
-    <header>
-      <a href="./index.html"><img src="./assets/back.svg" /></a>
-      <h1>Configurações</h1>
-    </header>
-
-    <div>
-      <h3>Dados de login:</h3>
-      <label>Email: </label>
-      <input type="email" id="set_email" /><br />
-      <label>Senha: </label>
-      <input type="password" id="set_password" /><br />
-      <button id="submit_set_login">login</button>
-    </div>
-    <br />
-    <div>
-      <h3>Forçar o envio de dados para o servidor:</h3>
-      <input type="date" id="sendData_data_inicio" />
-      <input type="date" id="sendData_data_fim" />
-      <button id="submit_sendData">enviar</button>
-    </div>
-    <br />
-    <div>
-      <h3>Dados do banco de dados local:</h3>
-      <label>User: </label>
-      <input type="text" id="set_db_user" /><br />
-      <label>Senha: </label>
-      <input type="text" id="set_db_password" /><br />
-      <label>Database: </label>
-      <input type="text" id="set_db_database" /><br />
-      <button id="submit_set_localdb">salvar</button>
-    </div>
-
-    <h2>INFO:</h2>
-    <div>
-      <div>
-        <h3>Login</h3>
-        <p id="info_email"></p>
-        <p>password: ******</p>
-      </div>
-      <div>
-        <h3>Banco de dados</h3>
-        <p id="info_user"></p>
-        <p>password: ******</p>
-        <p id="info_db"></p>
-      </div>
-      <div>
-        <p id="info_server_url"></p>
-      </div>
-      <button id="update_info">Atualizar</button>
-    </div>
-
-    <script src="config.js"></script>
-  </body>
-</html>
