--- style/article.css
@@ -1,12 +1,0 @@
-article {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: 5px;
-  padding: 10px;
-}
-
-.title_text {
-  font-size: 24px;
-  margin-left: 10px;
-}
