--- style/main.css
@@ -1,3 +1,0 @@
-main {
-    width: 100%;
-}
