--- style/section.css
@@ -1,113 +1,0 @@
-section {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: auto;
-  background-color: white;
-  gap: 10px;
-  margin-top: 20px;
-}
-section .block {
-  border: 0px solid black;
-  background-color: white;
-  filter: drop-shadow(4px 4px 4px #535353);
-  max-width: 1000px;
-}
-
-section .block_title {
-  font-size: 18px;
-  border-left: 5px solid var(--main-color);
-  padding-left: 8px;
-  /* font-weight: bold; */
-  /* color:#26b5f0; */
-}
-
-section .block_head {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px;
-}
-section .block_head_other {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  gap: 10px;
-  padding-left: 10px;
-}
-
-section .block_options {
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: end;
-  width: 100%;
-}
-
-section .block_options_date {
-  display: flex;
-  gap: 5px;
-}
-
-section .button_switch {
-  border: 1px solid black;
-  padding: 5px;
-}
-section .button_switch:hover {
-  cursor: pointer;
-}
-
-section .card_text {
-  border: 0px solid black;
-  background-color: white;
-  filter: drop-shadow(4px 4px 4px #535353);
-  padding-right: 20px;
-  padding-left: 20px;
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-section .card_section {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  grid-column: span 2;
-  column-gap: 30px;
-}
-
-section .card_text_info {
-  font-size: 16px;
-  font-weight: bold;
-  color: #26b5f0;
-  margin: 2px;
-}
-
-section .card_text_value {
-  font-size: 20px;
-  margin: 2px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  border: 1px solid;
-  padding: 10px;
-}
-
-thead {
-  background-color: #26b5f0;
-  font-weight: 700;
-  font-size: 18px;
-}
-th,
-td {
-  padding: 8px;
-  font-size: 16px;
-}
-
-@media (max-width: 1000px) {
-  section {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
