--- style/style.css
@@ -1,6 +1,0 @@
-@import url("./article.css");
-@import url("./geral.css");
-@import url("./section.css");
-@import url("./title_card.css");
-@import url("./main.css");
-
