--- style/title_card.css
@@ -1,52 +1,0 @@
-@import url("variable.css");
-
-.title {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-}
-
-.title_card {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 10px;
-  border: 0px solid black;
-  background-color: white;
-  filter: drop-shadow(4px 4px 4px var(--shadow-color));
-}
-
-.title_card .title {
-  margin: 0px;
-  padding: 10px;
-  border-left: 8px solid var(--main-color);
-  font-size: 24px;
-}
-
-.title_card .search_date {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 10px;
-}
-
-.title_card .search_date input {
-  margin-right: 10px;
-  font-size: 14px;
-  border: 1px solid black;
-  border-radius: 2px;
-  padding: 5px;
-}
-
-.title_card .search_date .search_button:hover {
-  cursor: pointer;
-}
-
-.title_card_content {
-  display: flex;
-}
