--- style/variable.css
@@ -1,10 +1,0 @@
-:root {
-  --main-color: #1cb2ef;
-  --secondary-color: #fcf2eb;
-  --secondary-color-selected: #f5ebe4;
-  --background-color: #f8f4f1;
-  --shadow-color: #686767;
-}
-body {
-  font-family: "Lato", serif;
-}
